Create Basic Arithmetic Operation
Write a C# program that can perform basic arithmetic functions. (Addition,
Subtraction, Multiplication, and Division). The user must be allowed to select which
arithmetic function

he/she wants to use, and then they will be prompted to enter only two numbers after
selecting the arithmetic function. Print the result afterwards, and then after printing
ask the user if he/she wants toperform another action or not. If yes, repeat the
program, if not terminate the program.